장바구니 조회 까지 추가

+++ prj2/src/main/java/com/be01/prj2/dto/cartDto/CartDto.java
@@ -13,8 +13,8 @@ public class CartDto {
 
     private Long cartId;
     private Long buyerId;
-    private Long totalQuantity;
+    private int totalQuantity;
     private CartStatus status;
-    private Long totalPrice;
+    private int totalPrice;
 
 }
